*** empty log message ***

+++ bundles/org.eclipse.swt/Eclipse SWT/gtk/org/eclipse/swt/widgets/Text.java
@@ -518,7 +518,7 @@ public int getCharCount () {
 	checkWidget ();
 	if ((style & SWT.SINGLE) != 0) {
 		int /*long*/ ptr = OS.gtk_entry_get_text (handle);
-		return OS.g_utf8_strlen (ptr, -1);
+		return (int)/*64*/OS.g_utf8_strlen (ptr, -1);
 	}
 	return OS.gtk_text_buffer_get_char_count (bufferHandle);
 }
